+++ SpatialGDK/Source/SpatialGDK/Private/Utils/RPCContainer.cpp
@@ -3,12 +3,11 @@
 #include "Utils/RPCContainer.h"
 
 #include "Schema/UnrealObjectRef.h"
+#include "SpatialGDKSettings.h"
 
 DEFINE_LOG_CATEGORY(LogRPCContainer);
 
 using namespace SpatialGDK;
-
-const double FRPCContainer::SECONDS_BEFORE_WARNING = 2.0;
 
 namespace
 {
@@ -77,7 +76,10 @@ namespace
 			*TimeDiff.ToString(),
 			*ERPCResultToString(ErrorInfo.ErrorCode));
 
-		if (TimeDiff.GetTotalSeconds() > FRPCContainer::SECONDS_BEFORE_WARNING)
+		const USpatialGDKSettings* SpatialGDKSettings = GetDefault<USpatialGDKSettings>();
+		check(SpatialGDKSettings != nullptr);
+
+		if (TimeDiff.GetTotalSeconds() > SpatialGDKSettings->GetSecondsBeforeWarning(ErrorInfo.ErrorCode))
 		{
 			UE_LOG(LogRPCContainer, Warning, TEXT("%s"), *OutputLog);
 		}